+++ sonar-plugin/sonar-javascript-plugin/src/test/java/org/sonar/plugins/javascript/analysis/JavaScriptEslintBasedSensorTest.java
@@ -164,7 +164,7 @@ class JavaScriptEslintBasedSensorTest {
     when(fileLinesContextFactory.createFor(any(InputFile.class))).thenReturn(fileLinesContext);
     analysisProcessor = new AnalysisProcessor(new DefaultNoSonarFilter(), fileLinesContextFactory);
     var analysisWarnings = new AnalysisWarningsWrapper();
-    var tsConfigCache = new TsConfigCache(bridgeServerMock);
+    var tsConfigCache = new TsConfigCacheImpl(bridgeServerMock);
 
     analysisWithProgram =
       new AnalysisWithProgram(bridgeServerMock, analysisProcessor, analysisWarnings);
