+++ sonar-plugin/sonar-javascript-plugin/src/test/java/org/sonar/plugins/javascript/analysis/JsTsSensorTest.java
@@ -175,7 +175,7 @@ class JsTsSensorTest {
     FileLinesContext fileLinesContext = mock(FileLinesContext.class);
     when(fileLinesContextFactory.createFor(any(InputFile.class))).thenReturn(fileLinesContext);
     processAnalysis = new AnalysisProcessor(new DefaultNoSonarFilter(), fileLinesContextFactory);
-    tsConfigCache = new TsConfigCache(bridgeServerMock);
+    tsConfigCache = new TsConfigCacheImpl(bridgeServerMock);
   }
 
   @Test
